Library

Library

Linear systems

Kalman.LinearHomogSystem — Type.

LinearHomogSystem

Linear dynamical model with linear observation scheme

source

StatsBase.sample — Function.

Y, X = sample(n, m, M::LinearHomogSystem)

Sample observations

n – number of observations per simulation m – number of independent simulations M – linear dynamic system and observation model

Y – observations (d2xnxm array) X – simulated processes (dxnxm array)

source

Filtering

Kalman.kalmanfilter — Function.

ll, xxf = kalmanfilter(yy::Matrix, M::LinearHomogSystem)

Kalman filter

yy -- d₂xn array
M -- linear model
ll -- marginal likelihood
xxf -- filtered process

source

ll, X = kalmanfilter{T}(Y::Array{T,3}, M::LinearHomogSystem)

Stack Kalman filter

Y -- array of m independent processes (dxnxm array)

X -- filtered processes (dxnxm array)
ll -- product marginal log likelihood

source

Kalman.observe! — Function.

observe!(s, x, P, t, y, M::LinearHomogSystem) -> t, y, H, R

source

Kalman.predict! — Function.

predict!(s, x, P, t, M::LinearHomogSystem) -> x, Ppred, Phi

source

Kalman.correct! — Function.

correct!(x, Ppred, y, H, R, SSM) -> x, P, yres, S, K

source

Kalman.evolve — Function.

evolve(s, x, P, t, M::LinearHomogSystem) -> x

source

Kalman.kalman_kernel — Function.

kalman_kernel(s, x, P, t, Y, SSM) -> t, x, P, Ppred, ll, K

Single Kalman filter step.

source

Smoothing

Kalman.kalmanrts — Function.

xxs, PP, PPpred, ll = kalmanrts(yy, xxs, M::LinearHomogSystem)

Rauch-Tung-Striebel smoother

yy -- d₂xn array or 
xxs -- empty dxn array or view
M --  linear dynamic system and observation model

xxs -- smoothed process
PP -- smoother variance
PPpred -- smoother prediction
ll -- filter likelihood

source

X = kalmanrts{T}(Y::Array{T,3}, M::LinearHomogSystem)

Stack Rauch-Tung-Striebel smoother, computes the marginal smoothed states, that is it computes the law $p(x_i | y_{1:n})$.

Y -- m independent processes (d₂xnxm array)

X -- smoothed processes (dxnxm array)

source

Iterators

Kalman.KalmanFilter — Type.

KalmanFilter(y, M)

Kalman filter as iterator, iterating over Gaussians representing the filtered distribution of x. Arguments y iterates over signal values.

source

Kalman.TimedKalmanFilter — Type.

TimedKalmanFilter(ty, M)

Kalman filter as iterator, iterating over Gaussians representing the filtered distribution of x. ty iterates over pairs (t, y) of signal time and signal value.

source

Tracking

Kalman.track — Function.

track(A::Matrix{Float64}, p, h = 10) -> p2, err

Track blurry lightsource by applying a window with half-width h at an approximate location p = (i,j) and find the average weighted location of points with high light intensity.

Gives an standard error estimate.

source