Library

Library

Linear systems

Kalman.LinearHomogSystem — Type.

LinearHomogSystem <: StateSpaceModel

LinearHomogSystem(x0, P0, Phi, b, Q, y, H, R)

Time homogeneous linear dynamical model with linear observation scheme corresponding to

$ x[k] = Φx[k−1] + b + w[k], w[k] ∼ N(0, Q) $

$ y[k] = Hx[k] + v[k], v[k] ∼ N(0, R) $

The argument y to the constructor is a dummy variable with the same type and size as the observations.

Example

x0 = [1., 0.]
P0 = eye(2)

Phi = [0.8 0.2; 0.0 0.8]
b = zeros(2)
Q = [0.1 0.0; 0.0 1.0]

y = [NaN]
H = [1.0 0.0]
R = eye(1)
M = LinearHomogSystem(x0, P0, Phi, b, Q, y, H, R)

source

StatsBase.sample — Function.

Y, X = sample(n, m, M::LinearHomogSystem) 

Sample observations generated by M.

n – number of observations per simulation

m – number of independent simulations

M – linear dynamic system and observation model

Y – observations (d2xnxm array)

X – simulated processes (dxnxm array)

source

Filtering

Kalman.kalmanfilter — Function.

ll, xxf = kalmanfilter(yy::Matrix, M::LinearHomogSystem)

Kalman filter

yy -- d₂xn array
M -- linear model
ll -- marginal likelihood
xxf -- filtered process

source

ll, X = kalmanfilter{T}(Y::Array{T,3}, M::LinearHomogSystem)

Stack Kalman filter

Y -- array of m independent processes (dxnxm array)

X -- filtered processes (dxnxm array)
ll -- product marginal log likelihood

source

Kalman.observe! — Function.

observe!(s, x, P, t, y, M) -> t, y, H, R

Obtain observation time t, observation y, observation matrix H and observation covariance H using model M. Except for certain use cases this is a no-op on arguments t and y (but needs to return t and y nevertheless.)

source

Kalman.predict! — Function.

predict!(s, x, P, t, M) -> x, Ppred, Phi

Computes predicted state x and predicted covariance of model M according the prediction equations

\[P_{pred} = \Phi P \Phi' + Q\]

Also returns evolution operator / Jacobian Phi.

source

Kalman.correct! — Function.

correct!(x, Ppred, y, H, R, M) -> x, P, yres, S, K

Perform the Kalman correction step using state space model M. Inputs are predicted state x and predicted covariance, as well as observation y and observation matrix H, observation covariance R.

Outputs filtered state x, filtered state covariance P, residual y as well as innovation covariance S and Kalman gain K.

source

Kalman.evolve — Function.

evolve(s, x, P, t, M) -> x

Evolve state x according to the state dynamics of model M without noise, usually

$ x \mapsto \Phi x + b. $

source

Kalman.kalman_kernel — Function.

kalman_kernel(s, x, P, t, Y, SSM) -> t, x, P, Ppred, ll, K

Single Kalman filter step consisting of a prediction step predict!, an observation step observe! and a correction step correct!. Return filtered covariance P and predicted Ppred

Computes and returns as well the log likelihood of the residual and the Kalman gain.

source

Smoothing

Kalman.kalmanrts — Function.

xxs, PP, PPpred, ll = kalmanrts(yy, xxs, M::LinearHomogSystem)

Rauch-Tung-Striebel smoother

yy -- d₂xn array or 
xxs -- empty dxn array or view
M --  linear dynamic system and observation model

xxs -- smoothed process
PP -- smoother variance
PPpred -- smoother prediction
ll -- filter likelihood

source

X = kalmanrts{T}(Y::Array{T,3}, M::LinearHomogSystem)

Stack Rauch-Tung-Striebel smoother, computes the marginal smoothed states, that is it computes the law $p(x_i | y_{1:n})$.

Y -- m independent processes (d₂xnxm array)

X -- smoothed processes (dxnxm array)

source

Iterators

Kalman.KalmanFilter — Type.

KalmanFilter(y, M)

Kalman filter as iterator, iterating over Gaussians representing the filtered distribution of x. Arguments y iterates over signal values.

Example

kf = KalmanFilter(Y, M) #
est = collect(kf)

source

Kalman.MappedKalmanFilter — Type.

MappedKalmanFilter(ty, M, f)

Kalman filter with model M as iterator, iterating over the result ret of ret = f(t, x, P, Ppred, ll, K). ty iterates over pairs (t, y) of signal time and signal value.

Eltype is determined by calling mappedreturntype(::typeof(f)) with fallback any.

Example

f(t, x, P, Ppred, ll, K) = ll
kf = MappedKalmanFilter(zip(Base.Iterators.countfrom(1), Y), M, f)
ll = sum(kf)

For performance, define:

import Kalman: mappedreturntype
Kalman.mappedreturntype(_, ::Type{typeof(f)}) = Float64

source

Tracking

Kalman.track — Function.

track(A::Matrix{Float64}, p, h = 10) -> p2, err

Track blurry lightsource by applying a window with half-width h at an approximate location p = (i,j) and find the average weighted location of points with high light intensity.

Gives an standard error estimate.

source